Publish same transform after a while
At the beginning of the execution the dynamic calculated transform is
broadcasted, after a short time interval has passed the last transform
is saved and broadcasted as a static transform since then.

The main issue at the moment is that in same parts of the 3D world
there is a small divergence between the position obtained from the
Visual Tracker and the position obtained from the Odometry.